More changes padded in
